pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "onepiece"
version = "0.1.0"
description = "Simple Flask Project"
requires-python = ">=3.8"
dependencies = [
    "Flask==3.0.0",
    "Flask-CORS==4.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "PyMySQL==1.1.0",
    "PyJWT==2.8.0",
    "cryptography==41.0.7",
]

[project.optional-dependencies]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["onepiece*"]
